+++ docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Latest
 
+## [0.1.2] 26 Feb, 2025
+
+### Added
+
+- Added Temperature and Humidity to the data collection process as well as the GUI.
+
+### Changed
+
+- Measurement unit details in the GUI.
+
 ## [0.1.1] 25 Feb, 2025
 
 - Added a PyQt6 GUI to the Auto Impedance Collection script